Move download tests to flaky
