+++ skyportal/tests/frontend/sources_and_followup_etc/test_source_list.py
@@ -9,8 +9,6 @@ from datetime import datetime, timedelta
 
 from baselayer.app.config import load_config
 from dateutil import parser
-from os.path import join as pjoin
-import random
 import numpy as np
 
 from selenium.webdriver.common.keys import Keys
@@ -688,97 +686,3 @@ def test_hr_diagram(
     driver.wait_for_xpath(f'//tr[@data-testid="groupSourceExpand_{source_id}"]')
 
     driver.wait_for_xpath(f'//div[@data-testid="hr_diagram_{source_id}"]')
-
-
-def test_download_sources(
-    driver, user, public_group, upload_data_token, annotation_token
-):
-    # generate a list of 20 source ids:
-    source_ids = [str(uuid.uuid4()) for i in range(20)]
-    origin = str(uuid.uuid4())
-    # post 20 sources:
-    for source_id in source_ids:
-        status, data = api(
-            "POST",
-            "sources",
-            data={
-                "id": source_id,
-                # random ra value
-                "ra": random.random() * 360 - 180,
-                "dec": random.random() * 180 - 90,
-                "redshift": 3,
-                "transient": False,
-                "ra_dis": 2.3,
-                "origin": origin,
-                "group_ids": [public_group.id],
-            },
-            token=upload_data_token,
-        )
-        assert status == 200
-
-    for source_id in source_ids:
-        status, data = api(
-            'POST',
-            f'sources/{source_id}/annotations',
-            data={
-                'origin': 'kowalski',
-                'data': {'offset_from_host_galaxy': 1.5},
-                'group_ids': [public_group.id],
-            },
-            token=annotation_token,
-        )
-        assert status == 200
-
-    for source_id in source_ids:
-        status, data = api(
-            'POST',
-            f'sources/{source_id}/annotations',
-            data={
-                'origin': 'other_origin',
-                'data': {
-                    'offset_from_host_galaxy': 1.5,
-                    'some_boolean': True,
-                },
-                'group_ids': [public_group.id],
-            },
-            token=annotation_token,
-        )
-        assert status == 200
-
-    driver.get(f"/become_user/{user.id}")
-    driver.get("/sources")
-
-    # Filter for origin
-    driver.click_xpath("//button[@data-testid='Filter Table-iconButton']")
-    alias_field = driver.wait_for_xpath(
-        "//*[@data-testid='origin-text']//input",
-    )
-    alias_field.send_keys(origin)
-    driver.click_xpath(
-        "//button[text()='Submit']",
-        scroll_parent=True,
-    )
-
-    # click the download button
-    driver.click_xpath('//button[@aria-label="Download CSV"]')
-
-    driver.wait_for_xpath('//*[contains(., "Downloading 20 sources")]')
-
-    driver.wait_for_xpath_to_disappear('//*[contains(., "Downloading 20 sources")]')
-
-    # check that the download has the right number of lines
-    fpath = str(os.path.abspath(pjoin(cfg['paths.downloads_folder'], 'sources.csv')))
-    try_count = 1
-    while not os.path.exists(fpath) and try_count <= 5:
-        try_count += 1
-        time.sleep(1)
-    assert os.path.exists(fpath)
-
-    try:
-        with open(fpath) as f:
-            lines = f.read()
-        assert len(lines.split('\n')) == 21
-        assert lines.split('\n')[1].find("kowalski;other_origin") != -1
-        assert lines.split('\n')[1].find("offset_from_host_galaxy;some_boolean") != -1
-    finally:
-        os.remove(fpath)
